requirements.txt
kafka-python==1.4.5
boto3==1.9.119
botocore==1.12.119
opencv-python==3.4.2.17
numpy==1.15.1
pymongo==3.7.2
azure-storage-blob==1.5.0